+++ src/components/Map/index.ts
@@ -1,3 +1,3 @@
 export { MapHandler } from './MapHandler'
-export { MapButtons } from './MapButtons'
+export { MapButtons } from './MapButtonGroups'
 export { MapPopupHandler } from './MapPopupHandler'
